Don't try to make the deploy group-writable.
The fast-remote-cache only really works when everyone who deploys,
deploys as the same user. This is because of the use of hard-links,
combined with the default deployment option in deploy:finalize_update
that tries to chmod the most recent release. Also, deploys go
significantly faster when you don't have to chmod an entire directory
tree. :)